# node-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Symptom Logger API",
    "version": "1.0.0",
    "description": "API documentation for the Symptom Logger API"
  },
  "servers": [
    {
      "url": "http://localhost:5200"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/user-symptom-log": {
      "post": {
        "summary": "Log a symptom for the user",
        "tags": [
          "UserSymptomLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symptoms": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "severity": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        "symptomStart": {
                          "type": "string",
                          "format": "date"
                        },
                        "symptomEnd": {
                          "type": "string",
                          "format": "date"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "countryCode": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Symptom log created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "summary": "Get all the symptoms logged by the user",
        "tags": [
          "UserSymptomLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of user symptom logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "symptoms": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "severity": {
                              "type": "integer",
                              "minimum": 1,
                              "maximum": 10
                            },
                            "symptomStart": {
                              "type": "string",
                              "format": "date"
                            },
                            "symptomEnd": {
                              "type": "string",
                              "format": "date"
                            },
                            "description": {
                              "type": "string"
                            }
                          }
                        }
                      },
                      "location": {
                        "type": "object",
                        "properties": {
                          "city": {
                            "type": "string"
                          },
                          "state": {
                            "type": "string"
                          },
                          "countryCode": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user-symptom-log/{id}": {
      "get": {
        "summary": "Get a specific symptom logged by the user",
        "tags": [
          "UserSymptomLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The symptom log ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Symptom log details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "symptoms": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "severity": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 10
                          },
                          "symptomStart": {
                            "type": "string",
                            "format": "date"
                          },
                          "symptomEnd": {
                            "type": "string",
                            "format": "date"
                          },
                          "description": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "location": {
                      "type": "object",
                      "properties": {
                        "city": {
                          "type": "string"
                        },
                        "state": {
                          "type": "string"
                        },
                        "countryCode": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Symptom log not found"
          }
        }
      }
    },
    "/user-symptom-log/{logId}": {
      "put": {
        "summary": "Update a specific symptom logged by the user",
        "tags": [
          "UserSymptomLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "logId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The symptom log ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symptoms": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "severity": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10
                        },
                        "symptomStart": {
                          "type": "string",
                          "format": "date"
                        },
                        "symptomEnd": {
                          "type": "string",
                          "format": "date"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "city": {
                        "type": "string"
                      },
                      "state": {
                        "type": "string"
                      },
                      "countryCode": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Symptom log updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Symptom log not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific symptom logged by the user",
        "tags": [
          "UserSymptomLog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "logId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The symptom log ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Symptom log deleted successfully"
          },
          "404": {
            "description": "Symptom log not found"
          }
        }
      }
    }
  },
...
}